Create HTML CSS CODE
This is a clean and simple login page created using HTML and CSS. It has fields for username and password, a login button, and a “Forgot Password?” link. The page is designed to be easy to use and looks neat with a centered form and soft colors. It works well on desktop browsers and can be a good starting point for any website’s login system.